Added allocator permissions on top of handle bit encoding
This adds support for using the permissions encoded in the low
2 bits of the allocator capability to confer allocation/free all
permission, supported with checks throughout the allocator APIs.
Scheme is as follows:

|bit 1|bit 2| Permissions            |
|-----|-----|------------------------|
|  1  |  1  | fully permissioned     |
|  1  |  0  | can free all, no alloc |
|  0  |  1  | can alloc, no free all |
|  0  |  0  | only ambient free perm |

If an allocator capability does not carry with it permission to
allocate, it also cannot be used to claim.

`malloc_capability_unseal` now takes an `int permissions` parameter
representing the required permissions, and returns `nullptr` if the
capability does not have all the specified permissions.

Notably, `heap_quota_remaining` is permission agnostic.
`allocate_sealed_unsealed` for now only checks `allocate` permissions,
but could be extended to use the third available bit in the handle to
further restrict the ability of a quota to be used to make sealed
allocations.

A new allocator.h header is introduced to provide the permissions API
(AllocatorPermission enum, allocator_permissions(), and
allocator_permissions_and()).