Allow large stacks.
This requires a CHERIoT-Platform/llvm-project#289 to work (without that,
the stacks are insufficiently aligned and the loader refuses to load
them).

With this, you can have stacks up to 64 KiB.  Tested with the PQC code
with a 60 KiB stack.  More than 64 KiB would require a small loader
change to not use the 16-bit size (and an associated linker change to
not refuse to link things with larger stacks).

The loader changes need careful review.  I believe I have checked the
important things, specifically:

 - Stacks are restored using the old (pre-adjustment) stack pointer.
 - Spills happen before the stack pointer is moved down.
 - The callee stack check happens with the size of the rounded-down
   stack.

All of these should be confirmed in review.